Updated and renamed jxl source
